fixed bugs, ready to make pr
