+++ my-app/src/Home.jsx
@@ -1,6 +1,7 @@
 import { ShoppingBag } from 'lucide-react'
 import React, { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom';
+import './App.css'
 
 const getAllListings = async () => {
   //example placeholders
